requirements.txt
astroid==2.11.3
certifi==2021.10.8
charset-normalizer==2.0.12
dill==0.3.4
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
platformdirs==2.5.2
pylint==2.13.7
python-gitlab==3.8.1
PyYAML==6.0
requests==2.27.1
requests-toolbelt==0.9.1
tomli==2.0.1
urllib3==1.26.9
watchdog==2.1.9
wrapt==1.14.0
